YatzyOperations.java became Yatzy.java
